chore: move autocomplete-service to common and fix imports
