+++ packages/ui/common/src/index.ts
@@ -33,8 +33,10 @@ export * from './lib/directives/index';
 export * from './lib/environments/environment';
 export * from './lib/components/delete-enity-dialog/delete-entity-dialog.component';
 export * from './lib/utils/beamer';
+export * from './lib/utils/insert-mention-operation';
 export * from './lib/directives/track-hover.directive';
 export * from './lib/service/templates.service';
 export * from './lib/directives/drag-drop.directive';
 export * from './lib/directives/element-ref.directive';
 export * from './lib/components/upload-file-control/upload-file-control.component';
+export * from './lib/service/builder-autocomplete-mentions-dropdown.service';
