+++ packages/ui/common/src/lib/utils/insert-mention-operation.ts
@@ -1,0 +1,9 @@
+export interface InsertMentionOperation {
+  insert: {
+    mention: {
+      value: string;
+      serverValue: string;
+      denotationChar: string;
+    };
+  };
+}
