+++ packages/ui/core/src/app/modules/flow-builder/page/flow-builder/collection-builder.component.ts
@@ -60,7 +60,7 @@ import {
   TemplateBlogNotificationComponent,
   BLOG_URL_TOKEN,
 } from '@activepieces/ui/feature-templates';
-import { BuilderAutocompleteMentionsDropdownService } from '@activepieces/ui/feature-builder-form-controls';
+import { BuilderAutocompleteMentionsDropdownService } from '@activepieces/ui/common';
 
 @Component({
   selector: 'app-collection-builder',
