+++ packages/ui/feature-builder-right-sidebar/src/lib/input-forms/loop-step-input-form/loop-step-input-form.component.ts
@@ -12,11 +12,8 @@ import {
 import { Observable, tap } from 'rxjs';
 import { ActionType } from '@activepieces/shared';
 import { LoopStepInputFormSchema } from '../input-forms-schema';
-import { fadeInUp400ms } from '@activepieces/ui/common';
-import {
-  InsertMentionOperation,
-  InterpolatingTextFormControlComponent,
-} from '@activepieces/ui/feature-builder-form-controls';
+import { fadeInUp400ms, InsertMentionOperation } from '@activepieces/ui/common';
+import { InterpolatingTextFormControlComponent } from '@activepieces/ui/feature-builder-form-controls';
 
 @Component({
   selector: 'app-loop-step-input-form',
